Lib

Documentation

count1s :: TuringMachine -> Integer

Function to count the ones on the tape. It is primaly good to be used for Busy Beavers and similar programs.

data Move

Implementation of the movement of the head of the Turing machine.

Constructors

L
L is used to move the head of the Turing machine to the left.

R
R is used to move the head of the Turing machine to the right.

run :: TuringMachine -> TuringMachine

Running the Turing machine without any output until the Turing machine halts. It returns the finished Turing machine.

runP :: TuringMachine -> IO TuringMachine

Version of run that prints the results after each step.

type State = Text

The State of the Turing machine is saved as a String.

step :: TuringMachine -> TuringMachine

Each step of the Turing machine is implemented in this function. It returns the resulting Turing machine.

stepP :: TuringMachine -> IO TuringMachine

Version of step that prints the result. It is intented for only a single step. For a complete run that should be printed use runP.

data Table

Implementation of the table of instructions. The States of the Turing machine are the keys of a Map. The values of this Map are the instructions. The instructions are another Map. The keys of that Map are the read Symbol. The values of that Map are a tuple which tells the Turing machine what to do. The first element of the tuple is the Symbol to write. The second element is the movement the head has to perform. The third element is the State in which the Turing machine switches afterwads.

Constructors

Table (Map State (Map Char (Char, Move, State)))

data Tape

Implementation of the tape of the Turing machine. The Symbols are stored in two Strings and a Char. The String on the left is the tape left of the head and the String on the right is the tape right of the head. The Char is the Symbol the head is over at a given moment. In this implementation of a Turing machine the head is part of the tape itself.

Constructors

Tape Text Char Text

toMove :: Text -> Maybe Move

data TuringMachine

The implementation of the Turing machine. The Turing machine has three values: the State it is in, the table of instructions and the tape.

Constructors

TM State Table Tape Integer