Main

Documentation

main :: IO ()

Main program that takes a TM description and text input file then prints the TuringTape if the TM description accepted the input, otherwise prints Nothing if the input was not accepted.