TuringMachine

Description

Defines all data structures used to create the TuringMachine such as the tape and transition functions.

Documentation

type State = String

Names of each state of the machine.

data Move

Represents a movement on the tape.

Constructors

L
Move left on the tape.

R
Move right on the tape.

data Cell

Data values that are in each cell of the tape.

Constructors

Cell String
A String representation of an input symbol.

Blank
For the blank symbol "_".

type TapeAlphabet = Set Cell

Uses a Set to get a unique alphabet for the TuringTape

type TMTape = TuringTape Cell

Instance of TuringTape using Cells as the values stored in the tape.

data TM

Instance of a TuringMachine.

Constructors

TM

Fields

currState :: State

Current State of the machine.

alphabet :: TapeAlphabet

Alphabet that this machine reads.

qa :: [State]

The accepting State.

qr :: [State]

The rejecting State.

delta :: [Transition]

List of Transitions which represent the transition table.

data Transition

A single case of the transition function of the machine.

Constructors

Transition

Fields

inState :: State

Input State, current state of the machine.

inSymbol :: Cell

Input symbol, current value the cursor of the tape is pointing at.

outState :: State

Output State for the next current state.

outSymbol :: Cell

Output symbol to write to the current cell of the tape.

outMove :: Move

Output Move for which direction to go along on the tape.

run :: TM -> TMTape -> Maybe TMTape

Runs a TM on a TuringTape.

step

Arguments

:: TM
Current iteration of the machine in its currState.

-> TMTape
Tape that the machine is reading through.

-> Maybe TMTape
Returns Nothing if the TuringMachine rejects. Just TMTape otherwise.

Each step of the TM, applying the Transition function based on the current symbol in the TuringTape and the current State