TuringMachineParser

Description

Parses a description given in the form: (Taken from specifications which can be found at https:/studres.cs.st-andrews.ac.ukCS3052PracticalsP01-turingmachines.pdf)

The first line is the word "states" followed by the number of states n.

Then there should be n lines, each containing the name of a state, possibly followed by a + for an accepting state, or - for a rejecting state.

Then a line with the world "alphabet" followed by the number of letters in the tape alphabet, followed by those letters, separated by spaces.

Then an arbitrary number of lines representing the transition table, each of which has the form <state1> <tapeinput> <state2> <tapeoutput> <move> where <move> is either L or R, <state1> and <state2> are mentioned in "states" <tapeinput> and <tapeoutput> are mentioned in "alphabet"

By convention, the first state mentioned is the start state. Not also that underscore (_) is used for the blank character, which does not need to be mentioned explicity in the alphabet.

This parser returns Maybe TM depending on whether the TuringMachine description was in the correct format.

Documentation

parseAlphabet

Arguments

:: [String]
List of strings in the format ["alphabet", n, ...] with n strings that denote a symbol in the alphabet.

-> Maybe TapeAlphabet
Returns Nothing if the list of strings is not in the specified format. Just TapeAlphabet otherwise.

Creates a Set which contains all Tape Alphabet.

parseStates

Arguments

:: Maybe TapeAlphabet
TapeAlphabet which the TM reads.

-> [String]
List of States the TM has, including the accept and rejecting States.

-> Maybe TM
Returns Nothing if the TapeAlphabet is Nothing. Just TM otherwise.

Creates a TM by parsing a list of States.

getCleanState

Arguments

:: String
State description.

-> String
Returns cleaned description that contains only the name of that State.

Sanitizes a State description by removing "+", "-" and " " from the String.

parseTransitions

Arguments

:: [String]
List of Transition descriptions

-> TM
TM to be updated. Any Transitions in the TM will be replaced.

-> TM
Returns updated TM.

Updates the TM with a list of Transitions parsed from the list of Strings.

createTransition

Arguments

:: [String]
Description of the transition. Must be in the format [<inputstate>, <inputsymbol>, <outputstate>, <outputsymbol>, <move>].

-> Transition
Returns the transition created by the description.

Creates a single Transition.

createCell

Arguments

:: String
String contents of the Cell

-> Cell
Returns a Cell. The Cell is Blank if the input String is an underscore (_)

Creates a single Cell of the TuringTape for the TuringMachine

createTM

Arguments

:: [String]
Description of the TM. Each element of the list is a line of the description.

-> Maybe TM
Returns Nothing if the description is in the wrong format. Just TM otherwise.

Creates a TM given a full description as described in the module description.