TuringTape

Description

The Tape for a TuringMachine is implemented as a ListZipper. This gives us O(1) operations on the list.

The cursor of the ListZipper is the current cell being read by the TuringMachine. The TuringMachine can then move the cursor left or right and write to the current cell.

Documentation

data TuringTape a

Constructors

T [a] a [a]

tapeLeft

Arguments

:: TuringTape a
Tape to move.

-> TuringTape a
Returns the moved tape.

Moves the tape cursor left.

tapeRight

Arguments

:: TuringTape a
Tape to move.

-> a
Default value if we have reached the end of the tape.

-> TuringTape a
Returns the moved tape.

Moves the tape cursor right.

currentCell :: TuringTape a -> a

Outputs the value of the current cell the cursor points at.

writeToCell

Arguments

:: TuringTape a
Tape to write over.

-> a
Value to write.

-> TuringTape a
Returns the tape with the value written over it.

Overwrites the value of the current cell the cursor points at.

initTape

Arguments

:: Ord a

=> [a]
List of values to store in the TuringTape.

-> Set a
TapeAlphabet where all values must be a part of.

-> Maybe (TuringTape a)
Returns Nothing if a value in the list does not belong to the TapeAlphabet. Just TuringTape otherwise.

Initialises a TuringTape.